Fix Authentication User

+++ src/main/java/com/tuananhdo/admin/repository/ProjectRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ProjectRepository extends CrudRepository<Project,Long> {
+public interface ProjectRepository extends CrudRepository<Project,Integer> {
 }
